Avoid stack overflow when saving multiple modified entities with the same key
Fixes #23043
