+++ test/EFCore.SqlServer.FunctionalTests/GraphUpdates/GraphUpdatesSqlServerClientNoActionTest.cs
@@ -1,0 +1,40 @@
+﻿// Licensed to the .NET Foundation under one or more agreements.
+// The .NET Foundation licenses this file to you under the MIT license.
+
+using System.Linq;
+using Microsoft.EntityFrameworkCore.Infrastructure;
+using Microsoft.EntityFrameworkCore.Storage;
+
+namespace Microsoft.EntityFrameworkCore
+{
+    public class GraphUpdatesSqlServerClientNoActionTest : GraphUpdatesSqlServerTestBase<GraphUpdatesSqlServerClientNoActionTest.SqlServerFixture>
+    {
+        public GraphUpdatesSqlServerClientNoActionTest(SqlServerFixture fixture)
+            : base(fixture)
+        {
+        }
+
+        protected override void UseTransaction(DatabaseFacade facade, IDbContextTransaction transaction)
+            => facade.UseTransaction(transaction.GetDbTransaction());
+
+        public class SqlServerFixture : GraphUpdatesSqlServerFixtureBase
+        {
+            public override bool ForceClientNoAction
+                => true;
+
+            protected override string StoreName { get; } = "GraphClientNoActionUpdatesTest";
+
+            protected override void OnModelCreating(ModelBuilder modelBuilder, DbContext context)
+            {
+                base.OnModelCreating(modelBuilder, context);
+
+                foreach (var foreignKey in modelBuilder.Model
+                    .GetEntityTypes()
+                    .SelectMany(e => e.GetDeclaredForeignKeys()))
+                {
+                    foreignKey.DeleteBehavior = DeleteBehavior.ClientNoAction;
+                }
+            }
+        }
+    }
+}
